requirements.txt
python-telegram-bot==20.8
openai==1.30.0
google-generativeai==0.3.0